@@ -18,18 +20,22 @@
     </licenses>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-        <fmt.maven.plugin.version>2.21.1</fmt.maven.plugin.version>
-        <enforcer.plugin.version>3.4.1</enforcer.plugin.version>
-        <compiler.plugin.version>3.11.0</compiler.plugin.version>
-        <asm.version>9.6</asm.version>
-        <surefire.plugin.version>3.2.2</surefire.plugin.version>
-        <jacoco.plugin.version>0.8.11</jacoco.plugin.version>
-        <fhir.parser.lib.version>1.0.7</fhir.parser.lib.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <version.maven.javadoc>3.11.2</version.maven.javadoc>
+        <version.maven-compiler-plugin>3.14.0</version.maven-compiler-plugin>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+        <fmt.maven.plugin.version>2.25</fmt.maven.plugin.version>
+        <enforcer.plugin.version>3.5.0</enforcer.plugin.version>
+        <asm.version>9.7.1</asm.version>
+        <surefire.plugin.version>3.5.2</surefire.plugin.version>
+        <jacoco.plugin.version>0.8.12</jacoco.plugin.version>
+        <fhir.parser.lib.version>2.4.1</fhir.parser.lib.version>
+        <junit.version>5.12.0</junit.version>
         <sonar.exclusions>
             **/de/gematik/demis/notification/builder/demis/fhir/notification/utils/DemisConstants.java
         </sonar.exclusions>
+        <argLine>-Dnet.bytebuddy.experimental=true</argLine>
     </properties>
 
     <dependencyManagement>
