@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.disease;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,166 +19,200 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
 
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.disease;
-
-import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.*;
+import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.CODE_SYSTEM_ACT_CODE;
+import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.STRUCTURE_DEFINITION_HOSPITALIZATION_NOTE;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.Utils.generateUuidString;
 import static java.util.Objects.requireNonNullElse;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
+import de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals.DateTimeTypeBuilder;
+import de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals.InitializableFhirObjectBuilder;
+import de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants;
+import de.gematik.demis.notification.builder.demis.fhir.notification.utils.Utils;
+import java.util.List;
 import lombok.EqualsAndHashCode;
 import lombok.Setter;
-import org.hl7.fhir.r4.model.*;
+import org.hl7.fhir.r4.model.CodeableConcept;
+import org.hl7.fhir.r4.model.Coding;
+import org.hl7.fhir.r4.model.DateTimeType;
+import org.hl7.fhir.r4.model.Encounter;
+import org.hl7.fhir.r4.model.Extension;
+import org.hl7.fhir.r4.model.Meta;
+import org.hl7.fhir.r4.model.Organization;
+import org.hl7.fhir.r4.model.Patient;
+import org.hl7.fhir.r4.model.Period;
+import org.hl7.fhir.r4.model.Reference;
+import org.hl7.fhir.r4.model.StringType;
 
 @Setter
 @EqualsAndHashCode
-public class EncounterDataBuilder {
+public class EncounterDataBuilder implements InitializableFhirObjectBuilder {
 
   private String id;
   private String profileUrl;
 
   private String hospitalizationNote;
-  private String hospitalizationRegionSystem;
-  private String hospitalizationRegionCode;
-  private String hospitalizationRegionDisplay;
-  private String hospitalizationRegionExtensionUrl;
+  private Coding hospitalizationRegion;
 
   private String status;
-  private String classSystem;
-  private String classCode;
-  private String classDisplay;
+  private Coding classification;
+  private Coding serviceType;
 
-  private String serviceTypeSystem;
-  private String serviceTypeCode;
-  private String serviceTypeDisplay;
+  private DateTimeType periodStart;
+  private DateTimeType periodEnd;
+  private Coding reason;
 
-  private String periodStart;
-  private String periodEnd;
+  private Patient notifiedPerson;
+  private Organization serviceProvider;
 
-  public Encounter buildExampleHospitalization(
-      Patient notifiedPerson, Organization serviceProvider) {
-
-    status = requireNonNullElse(status, "inprogress");
-
-    classSystem = requireNonNullElse(classSystem, CODE_SYSTEM_ACT_CODE);
-    classCode = requireNonNullElse(classCode, "IMP");
-    classDisplay = requireNonNullElse(classDisplay, "inpatient encounter");
-
-    profileUrl = requireNonNullElse(profileUrl, PROFILE_HOSPITALIZATION);
-
-    hospitalizationNote = requireNonNullElse(hospitalizationNote, "wichtige Zusatzinformation");
-
-    periodStart = requireNonNullElse(periodStart, "22.01.2022");
-
-    return buildHospitalization(notifiedPerson, serviceProvider);
+  private static Period getPeriod(Encounter encounter) {
+    var period = encounter.getPeriod();
+    if (period == null) {
+      period = new Period();
+      encounter.setPeriod(period);
+    }
+    return period;
   }
 
-  public Encounter buildHospitalization(Patient notifiedPerson, Organization serviceProvider) {
+  @Override
+  public Encounter build() {
     Encounter encounter = new Encounter();
-
-    id = requireNonNullElse(id, generateUuidString());
-    encounter.setId(id);
-
-    encounter.setSubject(new Reference(notifiedPerson));
-    encounter.setServiceProvider(new Reference(serviceProvider));
-
+    addId(encounter);
+    addSubject(encounter);
+    addServiceProvider(encounter);
     addClassCodingIfSet(encounter);
     addStatusIfSet(encounter);
     addProfileIfSet(encounter);
     addPeriodStartIfSet(encounter);
     addPeriodEndIfSet(encounter);
+    addReasonIfSet(encounter);
     addHospitalizationNotIfSet(encounter);
     addHospitalizationRegionIfSet(encounter);
     addServiceTypeCodingIfSet(encounter);
-
     return encounter;
   }
 
+  private void addReasonIfSet(Encounter encounter) {
+    if (this.reason != null) {
+      encounter.setReasonCode(List.of(new CodeableConcept().addCoding(this.reason)));
+    }
+  }
+
+  /**
+   * Sets:
+   *
+   * <ul>
+   *   <li>ID
+   *   <li>profile URL
+   *   <li>classification as inpatient encounter
+   * </ul>
+   *
+   * @return builder
+   */
+  @Override
+  public EncounterDataBuilder setDefaults() {
+    if (this.id == null) {
+      setId(Utils.generateUuidString());
+    }
+    if (this.profileUrl == null) {
+      setProfileUrl(DemisConstants.PROFILE_HOSPITALIZATION);
+    }
+    if (this.classification == null) {
+      setClassification(new Coding().setSystem(CODE_SYSTEM_ACT_CODE).setCode("IMP"));
+    }
+    return this;
+  }
+
+  public EncounterDataBuilder setPeriodStart(String start) {
+    setPeriodStart(new DateTimeTypeBuilder().setText(start).build());
+    return this;
+  }
+
+  public EncounterDataBuilder setPeriodEnd(String end) {
+    setPeriodEnd(new DateTimeTypeBuilder().setText(end).build());
+    return this;
+  }
+
+  public EncounterDataBuilder setPeriodStart(DateTimeType start) {
+    this.periodStart = start;
+    return this;
+  }
+
+  public EncounterDataBuilder setPeriodEnd(DateTimeType end) {
+    this.periodEnd = end;
+    return this;
+  }
+
+  private void addSubject(Encounter encounter) {
+    if (this.notifiedPerson != null) {
+      encounter.setSubject(new Reference(this.notifiedPerson));
+    }
+  }
+
+  private void addServiceProvider(Encounter encounter) {
+    if (this.serviceProvider != null) {
+      encounter.setServiceProvider(new Reference(this.serviceProvider));
+    }
+  }
+
+  private void addId(Encounter encounter) {
+    this.id = requireNonNullElse(this.id, generateUuidString());
+    encounter.setId(this.id);
+  }
+
   private void addServiceTypeCodingIfSet(Encounter encounter) {
-    if (isNotBlank(serviceTypeSystem)
-        || isNotBlank(serviceTypeCode)
-        || isNotBlank(serviceTypeDisplay)) {
-      Coding serviceTypeCoding = new Coding(serviceTypeSystem, serviceTypeCode, serviceTypeDisplay);
-      encounter.setServiceType(new CodeableConcept(serviceTypeCoding));
+    if (this.serviceType != null) {
+      encounter.setServiceType(new CodeableConcept(this.serviceType));
     }
   }
 
   private void addHospitalizationRegionIfSet(Encounter encounter) {
-    if (isNotBlank(hospitalizationRegionSystem)
-        || isNotBlank(hospitalizationRegionCode)
-        || isNotBlank(hospitalizationRegionDisplay)) {
-      hospitalizationRegionExtensionUrl =
-          requireNonNullElse(
-              hospitalizationRegionExtensionUrl, STRUCTURE_DEFINITION_HOSPITALIZATION_REGION);
-      Extension t = new Extension();
-      t.setUrl(hospitalizationRegionExtensionUrl);
-      Coding coding =
-          new Coding(
-              hospitalizationRegionSystem, hospitalizationRegionCode, hospitalizationRegionDisplay);
-      CodeableConcept codeableConcept = new CodeableConcept(coding);
-      t.setValue(codeableConcept);
-      encounter.addExtension(t);
+    if (this.hospitalizationRegion != null) {
+      Extension region = new Extension();
+      region.setUrl(DemisConstants.STRUCTURE_DEFINITION_HOSPITALIZATION_REGION);
+      region.setValue(new CodeableConcept(this.hospitalizationRegion));
+      encounter.addExtension(region);
     }
   }
 
   private void addHospitalizationNotIfSet(Encounter encounter) {
-    if (isNotBlank(hospitalizationNote)) {
+    if (isNotBlank(this.hospitalizationNote)) {
       encounter.addExtension(
           new Extension(STRUCTURE_DEFINITION_HOSPITALIZATION_NOTE)
-              .setValue(new StringType(hospitalizationNote)));
-    }
-  }
-
-  private void addPeriodEndIfSet(Encounter encounter) {
-    if (this.periodEnd != null) {
-      Date value = parse(this.periodEnd);
-      if (encounter.getPeriod() == null) {
-        encounter.setPeriod(new Period().setEnd(value));
-      } else {
-        encounter.getPeriod().setEnd(value);
-      }
-    }
-  }
-
-  private static Date parse(String date) {
-    try {
-      return new SimpleDateFormat("dd.MM.yyyy").parse(date);
-    } catch (ParseException e) {
-      throw new IllegalArgumentException("Failed to parse date: " + date, e);
+              .setValue(new StringType(this.hospitalizationNote)));
     }
   }
 
   private void addPeriodStartIfSet(Encounter encounter) {
     if (this.periodStart != null) {
-      Date value = parse(this.periodStart);
-      if (encounter.getPeriod() == null) {
-        encounter.setPeriod(new Period().setStart(value));
-      } else {
-        encounter.getPeriod().setStart(value);
-      }
+      getPeriod(encounter).setStartElement(this.periodStart);
+    }
+  }
+
+  private void addPeriodEndIfSet(Encounter encounter) {
+    if (this.periodEnd != null) {
+      getPeriod(encounter).setEndElement(this.periodEnd);
     }
   }
 
   private void addProfileIfSet(Encounter encounter) {
-    if (isNotBlank(profileUrl)) {
-      encounter.setMeta(new Meta().addProfile(profileUrl));
+    if (isNotBlank(this.profileUrl)) {
+      encounter.setMeta(new Meta().addProfile(this.profileUrl));
     }
   }
 
   private void addStatusIfSet(Encounter encounter) {
-    if (isNotBlank(status)) {
-      encounter.setStatus(Encounter.EncounterStatus.valueOf(status.toUpperCase()));
+    if (isNotBlank(this.status)) {
+      encounter.setStatus(Encounter.EncounterStatus.fromCode(this.status.toLowerCase()));
     }
   }
 
   private void addClassCodingIfSet(Encounter encounter) {
-    if (isNotBlank(classSystem) || isNotBlank(classCode) || isNotBlank(classDisplay)) {
-      encounter.setClass_(new Coding(classSystem, classCode, classDisplay));
+    if (this.classification != null) {
+      encounter.setClass_(this.classification);
     }
   }
 }
