@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.disease;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,169 +19,203 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
 
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.disease;
-
-import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.BundleUtils.addEntry;
-import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.NAMING_SYSTEM_NOTIFICATION_BUNDLE_ID;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.PROFILE_NOTIFICATION_BUNDLE_DISEASE;
-import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.Utils.generateUuidString;
-import static java.util.Objects.requireNonNullElse;
-import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
-import de.gematik.demis.notification.builder.demis.fhir.notification.builder.NotifierDataBuilder;
-import de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.NotifiedPersonDataBuilder;
-import de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.disease.questionnaire.DiseaseInformationCVDDDataBuilder;
-import de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.disease.questionnaire.DiseaseInformationCommonDataBuilder;
-import de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.disease.questionnaire.ExampleQuestionnaireCVVDData;
-import de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.disease.questionnaire.ExampleQuestionnaireCommonData;
-import java.util.Collections;
-import java.util.Date;
+import de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals.BundleDataBuilder;
+import de.gematik.demis.notification.builder.demis.fhir.notification.utils.Utils;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+import lombok.Getter;
 import lombok.Setter;
-import org.hl7.fhir.r4.model.*;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.hl7.fhir.r4.model.CanonicalType;
+import org.hl7.fhir.r4.model.Composition;
+import org.hl7.fhir.r4.model.Condition;
+import org.hl7.fhir.r4.model.DateTimeType;
+import org.hl7.fhir.r4.model.Encounter;
+import org.hl7.fhir.r4.model.Immunization;
+import org.hl7.fhir.r4.model.InstantType;
+import org.hl7.fhir.r4.model.Meta;
+import org.hl7.fhir.r4.model.Organization;
+import org.hl7.fhir.r4.model.Patient;
+import org.hl7.fhir.r4.model.PractitionerRole;
+import org.hl7.fhir.r4.model.QuestionnaireResponse;
+import org.hl7.fhir.r4.model.Resource;
 
+/**
+ * Builder for a FHIR bundle containing a disease notification. Special feature: <code>
+ * createComposition()</code> creates a composition based on the configured data.
+ */
 @Setter
-public class NotificationBundleDiseaseDataBuilder {
+@Slf4j
+public class NotificationBundleDiseaseDataBuilder extends BundleDataBuilder {
 
-  private String id;
-  private String profileUrl;
-  private String identifierString;
-  private String identifierSystem;
-  private String type;
-  private Date timestamp;
+  static final String CONDITION_PROFILE_PREFIX =
+      "https://demis.rki.de/fhir/StructureDefinition/Disease";
+
+  private final List<Resource> hospitalizations = new ArrayList<>();
+  private final List<Resource> immunizations = new ArrayList<>();
+  private final List<Resource> organizations = new ArrayList<>();
 
   private Patient notifiedPerson;
+  @Getter private PractitionerRole notifierRole;
 
-  private PractitionerRole notifierRole;
-
+  /** The FHIR bundle composition entry will be generated automatically if you don't set it. */
   private Composition notificationDisease;
 
   private Condition disease;
+  private QuestionnaireResponse commonInformation;
+  private QuestionnaireResponse specificInformation;
 
-  private QuestionnaireResponse diseaseInformationCommon;
-
-  private QuestionnaireResponse diseaseInformationSpecific;
-
-  private List<Resource> encounterList = Collections.emptyList();
-  private List<Resource> immunizationList = Collections.emptyList();
-  private List<Resource> encounterOrganizationList = Collections.emptyList();
-
-  private List<QuestionnaireResponse.QuestionnaireResponseItemComponent>
-      questionnaireResponsesForCommonQuesitons = Collections.emptyList();
-  private List<QuestionnaireResponse.QuestionnaireResponseItemComponent>
-      questionnaireResponsesForSpecificQuestions = Collections.emptyList();
-
-  /**
-   * @deprecated no example code data will be provided in the future. please create your own
-   *     testdata in your project.
-   * @return
-   */
-  @Deprecated(since = "1.2.1")
-  public Bundle buildExampleCVDDDiseaseBundle() {
-
-    identifierString = requireNonNullElse(identifierSystem, generateUuidString());
-    identifierSystem = requireNonNullElse(identifierSystem, NAMING_SYSTEM_NOTIFICATION_BUNDLE_ID);
-
-    timestamp = new Date();
-    type = requireNonNullElse(type, "document");
-
-    notifiedPerson =
-        requireNonNullElse(
-            notifiedPerson, new NotifiedPersonDataBuilder().buildExampleNotifiedPerson());
-    notifierRole =
-        requireNonNullElse(
-            notifierRole, new NotifierDataBuilder().buildLaboratoryExampleNotifierDataHospital());
-
-    disease =
-        requireNonNullElse(
-            disease, new DiseaseDataBuilder().buildCVDDexampleDisease(notifiedPerson));
-
-    if (encounterList.isEmpty()) {
-      encounterList =
-          Collections.singletonList(
-              new EncounterDataBuilder()
-                  .buildExampleHospitalization(
-                      notifiedPerson, (Organization) notifierRole.getOrganization().getResource()));
+  private static String validSuffix(String disease) {
+    if ((StringUtils.length(disease) != 4) || !StringUtils.isAlpha(disease)) {
+      throw new IllegalArgumentException("Not a four letter disease identifier: " + disease);
     }
-
-    diseaseInformationCommon =
-        requireNonNullElse(
-            diseaseInformationCommon,
-            new DiseaseInformationCommonDataBuilder()
-                .buildExampleDiseaseInformationCommonData(
-                    notifiedPerson,
-                    ExampleQuestionnaireCommonData.getCommonQuestionnaire(
-                        (Organization) notifierRole.getOrganization().getResource(),
-                        encounterList)));
-
-    if (immunizationList.isEmpty()) {
-      immunizationList =
-          Collections.singletonList(
-              new ImmunizationDataBuilder().buildExampleImmunization(notifiedPerson));
-    }
-
-    diseaseInformationSpecific =
-        requireNonNullElse(
-            diseaseInformationSpecific,
-            new DiseaseInformationCVDDDataBuilder()
-                .buildExampleCVDDInformationSpecificData(
-                    notifiedPerson,
-                    ExampleQuestionnaireCVVDData.getCVDDQuestionnaire(immunizationList)));
-
-    notificationDisease =
-        requireNonNullElse(
-            notificationDisease,
-            new NotificationDiseaseDataBuilder()
-                .buildExampleCVDDNotificationDisease(
-                    notifiedPerson,
-                    notifierRole,
-                    disease,
-                    diseaseInformationCommon,
-                    diseaseInformationSpecific));
-
-    return buildDiseaseBundle();
+    return disease.toUpperCase();
   }
 
-  public Bundle buildDiseaseBundle() {
-    Bundle bundle = new Bundle();
+  /**
+   * Create disease specific URL.
+   *
+   * @param url URL
+   * @param disease disease category code like: <code>cvdd</code>
+   * @return URL
+   */
+  public static String createDiseaseSpecificUrl(String url, String disease) {
+    return url + validSuffix(disease);
+  }
 
-    id = requireNonNullElse(id, generateUuidString());
-    bundle.setId(id);
-    profileUrl = requireNonNullElse(profileUrl, PROFILE_NOTIFICATION_BUNDLE_DISEASE);
-    bundle.setMeta(new Meta().addProfile(profileUrl));
+  public final NotificationBundleDiseaseDataBuilder addImmunization(Immunization immunization) {
+    this.immunizations.add(immunization);
+    return this;
+  }
 
-    Identifier identifier = new Identifier();
-    if (isNotBlank(identifierString)) {
-      identifier.setValue(identifierString);
+  @Override
+  public NotificationBundleDiseaseDataBuilder setDefaults() {
+    super.setDefaults();
+    return this;
+  }
+
+  public final NotificationBundleDiseaseDataBuilder setImmunizations(
+      List<Immunization> immunizations) {
+    this.immunizations.clear();
+    this.immunizations.addAll(immunizations);
+    return this;
+  }
+
+  public final NotificationBundleDiseaseDataBuilder addHospitalization(Encounter hospitalization) {
+    this.hospitalizations.add(hospitalization);
+    return this;
+  }
+
+  public final NotificationBundleDiseaseDataBuilder setHospitalizations(
+      List<Encounter> hospitalizations) {
+    this.hospitalizations.clear();
+    this.hospitalizations.addAll(hospitalizations);
+    return this;
+  }
+
+  public final NotificationBundleDiseaseDataBuilder addOrganization(Organization organization) {
+    this.organizations.add(organization);
+    return this;
+  }
+
+  public final NotificationBundleDiseaseDataBuilder setOrganizations(
+      List<Organization> organizations) {
+    this.organizations.clear();
+    this.organizations.addAll(organizations);
+    return this;
+  }
+
+  @Override
+  protected void addEntries() {
+    addComposition();
+    addEntry(this.notifiedPerson);
+    addEntry(this.disease);
+    addEntryOfPractitionerRole(this.notifierRole);
+    addQuestionnaireResponses();
+  }
+
+  @Override
+  protected String getDefaultProfileUrl() {
+    return PROFILE_NOTIFICATION_BUNDLE_DISEASE;
+  }
+
+  /**
+   * Create composition based on configured data.
+   *
+   * @return composition builder
+   */
+  public NotificationDiseaseDataBuilder createComposition() {
+    log.debug("Disease notification bundle builder generates composition.");
+    NotificationDiseaseDataBuilder composition = new NotificationDiseaseDataBuilder();
+    composition.setDefaults();
+    composition.setProfileUrlByDisease(getCategory());
+    DateTimeType date = new DateTimeType();
+    date.setValue(Utils.getCurrentDate(), InstantType.DEFAULT_PRECISION);
+    composition.setDate(date);
+    composition.setNotifiedPerson(this.notifiedPerson);
+    composition.setNotifierRole(this.notifierRole);
+    composition.setDisease(this.disease);
+    composition.setCommonQuestionnaireResponse(this.commonInformation);
+    if (this.specificInformation != null) {
+      composition.setSpecificQuestionnaireResponse(this.specificInformation);
     }
-    if (isNotBlank(identifierSystem)) {
-      identifier.setSystem(identifierSystem);
-    }
-    bundle.setIdentifier(identifier);
+    return composition;
+  }
 
-    if (isNotBlank(type)) {
-      bundle.setType(Bundle.BundleType.valueOf(type.toUpperCase()));
+  private String getCategory() {
+    if (this.disease.hasMeta()) {
+      Meta meta = this.disease.getMeta();
+      if (meta.hasProfile()) {
+        final Optional<String> category =
+            meta.getProfile().stream()
+                .map(CanonicalType::getValue)
+                .filter(p -> StringUtils.startsWith(p, CONDITION_PROFILE_PREFIX))
+                .map(p -> p.substring(p.length() - 4))
+                .findFirst();
+        if (category.isPresent()) {
+          return category.get();
+        }
+      }
     }
+    throw new IllegalArgumentException("disease FHIR condition resource has no profile URL");
+  }
 
-    if (timestamp != null) {
-      bundle.setTimestamp(timestamp);
+  private void addComposition() {
+    if (this.notificationDisease != null) {
+      addEntry(this.notificationDisease);
     }
+  }
 
-    addEntry(bundle, notificationDisease);
-    addEntry(bundle, notifiedPerson);
-    addEntry(bundle, disease);
-    addEntry(bundle, notifierRole);
-    if (notifierRole.getOrganization() != null) {
-      addEntry(bundle, (Organization) notifierRole.getOrganization().getResource());
+  private void addQuestionnaireResponses() {
+    addCommonInformationQuestionnaireResponse();
+    addInformationQuestionnaireResponse();
+  }
+
+  private void addInformationQuestionnaireResponse() {
+    if (this.specificInformation != null) {
+      addImmunizations();
+      addEntry(this.specificInformation);
     }
-    addEntry(bundle, encounterList);
-    addEntry(bundle, encounterOrganizationList);
-    addEntry(bundle, diseaseInformationCommon);
-    addEntry(bundle, immunizationList);
-    addEntry(bundle, diseaseInformationSpecific);
+  }
 
-    return bundle;
+  private void addCommonInformationQuestionnaireResponse() {
+    addEncounters();
+    addEntry(this.commonInformation);
+  }
+
+  private void addImmunizations() {
+    addEntries(this.immunizations);
+  }
+
+  private void addEncounters() {
+    addEntries(this.organizations);
+    addEntries(this.hospitalizations);
   }
 }
