@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.laboratory;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,22 +19,32 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
-
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.laboratory;
 
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.LABORAOTRY_CATEGORY_CODE;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.LABORATORY_CATEGORY_DISPLAY;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.LOINC_ORG_SYSTEM;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.NOTIFICATION_ID_SYSTEM;
+import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.PROFILE_NOTIFICATION_LABORATORY;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.Utils.generateUuidString;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.Utils.getCurrentDate;
 import static java.util.Objects.requireNonNullElse;
 
 import de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals.CompositionBuilder;
 import de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants;
+import de.gematik.demis.notification.builder.demis.fhir.notification.utils.ReferenceUtils;
+import javax.annotation.CheckForNull;
 import lombok.Setter;
-import org.hl7.fhir.r4.model.*;
+import org.hl7.fhir.r4.model.CodeableConcept;
+import org.hl7.fhir.r4.model.Coding;
+import org.hl7.fhir.r4.model.Composition;
+import org.hl7.fhir.r4.model.DiagnosticReport;
+import org.hl7.fhir.r4.model.Identifier;
+import org.hl7.fhir.r4.model.Meta;
+import org.hl7.fhir.r4.model.Patient;
+import org.hl7.fhir.r4.model.PractitionerRole;
+import org.hl7.fhir.r4.model.Reference;
 
 @Setter
 public class NotificationLaboratoryDataBuilder
@@ -39,18 +54,18 @@ public class NotificationLaboratoryDataBuilder
   private String sectionComponentCode;
   private String sectionComponentDisplay;
 
-  private DiagnosticReport laboratoryReport;
+  @CheckForNull private DiagnosticReport laboratoryReport;
 
   public NotificationLaboratoryDataBuilder setDefault() {
 
-    metaUrl = DemisConstants.PROFILE_NOTIFICATION_LABORATORY;
-    identifierSystem = NOTIFICATION_ID_SYSTEM;
-    sectionComponentSystem = LOINC_ORG_SYSTEM;
-    codeAndCategorySystem = LOINC_ORG_SYSTEM;
-    codeAndCategoryCode = LABORAOTRY_CATEGORY_CODE;
-    codeAndCategoryDisplay = LABORATORY_CATEGORY_DISPLAY;
+    setIdentifierSystem(NOTIFICATION_ID_SYSTEM);
+    setSectionComponentSystem(LOINC_ORG_SYSTEM);
+    setCodeAndCategorySystem(LOINC_ORG_SYSTEM);
+    setCodeAndCategoryCode(LABORAOTRY_CATEGORY_CODE);
+    setCodeAndCategoryDisplay(LABORATORY_CATEGORY_DISPLAY);
+    setTitle("Erregernachweismeldung");
 
-    super.setDefaultdData();
+    super.setDefaultData();
     return this;
   }
 
@@ -58,32 +73,58 @@ public class NotificationLaboratoryDataBuilder
   public Composition build() {
     Composition newComposition = super.build();
 
+    newComposition.setTitle(getTitle());
     Identifier compositionIdentifier =
-        new Identifier().setSystem(identifierSystem).setValue(identifierValue);
+        new Identifier().setSystem(getIdentifierSystem()).setValue(getIdentifierValue());
     newComposition.setIdentifier(compositionIdentifier);
 
-    Reference laboratoryReportReference = new Reference(laboratoryReport);
-    Composition.SectionComponent sectionComponent = new Composition.SectionComponent();
-    sectionComponent.addEntry(laboratoryReportReference);
-    sectionComponent.setCode(
-        new CodeableConcept(
-            new Coding(sectionComponentSystem, sectionComponentCode, sectionComponentDisplay)));
-    newComposition.addSection(sectionComponent);
+    if (laboratoryReport != null) {
+      Composition.SectionComponent sectionComponent = new Composition.SectionComponent();
+      sectionComponent.addEntry(ReferenceUtils.internalReference(laboratoryReport));
+      sectionComponent.setCode(
+          new CodeableConcept(
+              new Coding(sectionComponentSystem, sectionComponentCode, sectionComponentDisplay)));
+      newComposition.addSection(sectionComponent);
+    }
 
+    String relatesTo = getRelatesTo();
     if (relatesTo != null) {
       Reference reference = new Reference();
       Identifier identifier =
           new Identifier().setSystem(NOTIFICATION_ID_SYSTEM).setValue(relatesTo);
       reference.setIdentifier(identifier);
-      reference.setType("Composition");
+      reference.setType(getRelatesToReferenceType());
       Composition.CompositionRelatesToComponent compositionRelatesToComponent =
           new Composition.CompositionRelatesToComponent();
-      compositionRelatesToComponent.setCode(Composition.DocumentRelationshipType.APPENDS);
+      compositionRelatesToComponent.setCode(getRelatesToCode());
       compositionRelatesToComponent.setTarget(reference);
       newComposition.addRelatesTo(compositionRelatesToComponent);
     }
 
     return newComposition;
+  }
+
+  protected NotificationLaboratoryDataBuilder setTypeCode(final Coding type) {
+    return this.setTypeCode(type.getCode())
+        .setTypeDisplay(type.getDisplay())
+        .setTypeSystem(type.getSystem());
+  }
+
+  protected NotificationLaboratoryDataBuilder setCategoryCode(final Coding category) {
+    return this.setCodeAndCategoryCode(category.getCode())
+        .setCodeAndCategoryDisplay(category.getDisplay())
+        .setCodeAndCategorySystem(category.getSystem());
+  }
+
+  protected NotificationLaboratoryDataBuilder setSectionCode(final Coding section) {
+    return this.setSectionComponentCode(section.getCode())
+        .setSectionComponentDisplay(section.getDisplay())
+        .setSectionComponentSystem(section.getSystem());
+  }
+
+  @Override
+  protected String getDefaultProfileUrl() {
+    return PROFILE_NOTIFICATION_LABORATORY;
   }
 
   /**
@@ -93,27 +134,28 @@ public class NotificationLaboratoryDataBuilder
   public Composition buildNotificationLaboratory(
       Patient notifiedPerson, PractitionerRole notifierRole, DiagnosticReport laboratoryReport) {
     Composition newComposition = new Composition();
-    newComposition.setId(notificationId);
-    newComposition.setStatus(compositionStatus);
+    newComposition.setId(getNotificationId());
+    newComposition.setStatus(getCompositionStatus());
 
     Composition.SectionComponent sectionComponent =
         setMandatoryFields(notifiedPerson, notifierRole, laboratoryReport, newComposition);
 
     Coding codeAndCategroyCoding =
-        new Coding(codeAndCategorySystem, codeAndCategoryCode, codeAndCategoryDisplay);
+        new Coding(
+            getCodeAndCategorySystem(), getCodeAndCategoryCode(), getCodeAndCategoryDisplay());
     CodeableConcept categoryCodeableConcept = new CodeableConcept(codeAndCategroyCoding);
     newComposition.addCategory(categoryCodeableConcept);
     sectionComponent.setCode(categoryCodeableConcept);
 
     newComposition.setMeta(new Meta().addProfile(DemisConstants.PROFILE_NOTIFICATION_LABORATORY));
-    newComposition.setDate(date);
-    newComposition.setTitle(title);
+    newComposition.setDate(getDate());
+    newComposition.setTitle(getTitle());
 
     Identifier compositionIdentifier =
-        new Identifier().setSystem(identifierSystem).setValue(identifierValue);
+        new Identifier().setSystem(getIdentifierSystem()).setValue(getIdentifierValue());
     newComposition.setIdentifier(compositionIdentifier);
 
-    Coding typeCoding = new Coding(typeSystem, typeCode, typeDisplay);
+    Coding typeCoding = new Coding(getTypeSystem(), getTypeCode(), getTypeDisplay());
     CodeableConcept typeCodeableConcept = new CodeableConcept(typeCoding);
     newComposition.setType(typeCodeableConcept);
 
@@ -126,13 +168,13 @@ public class NotificationLaboratoryDataBuilder
   @Deprecated(since = "1.2.1")
   public Composition buildExampleComposition(
       Patient notifiedPerson, PractitionerRole notifierRole, DiagnosticReport laboratoryReport) {
-    notificationId = requireNonNullElse(notificationId, generateUuidString());
+    setNotificationId(requireNonNullElse(getNotificationId(), generateUuidString()));
     setExampleCompositionStatus();
     setExampleType();
     setExampleCodeAndCategory();
     setExampleIdentifier();
-    date = requireNonNullElse(date, getCurrentDate());
-    title = requireNonNullElse(title, "Erregernachweismeldung");
+    setDate(requireNonNullElse(getDate(), getCurrentDate()));
+    setTitle(requireNonNullElse(getTitle(), "Erregernachweismeldung"));
     return buildNotificationLaboratory(notifiedPerson, notifierRole, laboratoryReport);
   }
 
@@ -150,8 +192,8 @@ public class NotificationLaboratoryDataBuilder
    */
   @Deprecated(since = "1.2.1")
   private NotificationLaboratoryDataBuilder addIdentifier(String value, String system) {
-    identifierSystem = system;
-    identifierValue = value;
+    setIdentifierSystem(system);
+    setIdentifierValue(value);
     return this;
   }
 
@@ -160,10 +202,10 @@ public class NotificationLaboratoryDataBuilder
    */
   @Deprecated(since = "1.2.1")
   private void setExampleIdentifier() {
-    identifierSystem =
+    setIdentifierSystem(
         requireNonNullElse(
-            identifierSystem, "https://demis.rki.de/fhir/NamingSystem/NotificationId");
-    identifierValue = requireNonNullElse(identifierValue, generateUuidString());
+            getIdentifierSystem(), "https://demis.rki.de/fhir/NamingSystem/NotificationId"));
+    setIdentifierValue(requireNonNullElse(getIdentifierValue(), generateUuidString()));
   }
 
   /**
@@ -171,9 +213,9 @@ public class NotificationLaboratoryDataBuilder
    */
   @Deprecated(since = "1.2.1")
   private void setExampleCodeAndCategory() {
-    codeAndCategorySystem = requireNonNullElse(codeAndCategorySystem, "http://loinc.org");
-    codeAndCategoryCode = requireNonNullElse(codeAndCategoryCode, "11502-2");
-    codeAndCategoryDisplay = requireNonNullElse(codeAndCategoryDisplay, "Laboratory report");
+    setCodeAndCategorySystem(requireNonNullElse(getCodeAndCategorySystem(), "http://loinc.org"));
+    setCodeAndCategoryCode(requireNonNullElse(getCodeAndCategoryCode(), "11502-2"));
+    setCodeAndCategoryDisplay(requireNonNullElse(getCodeAndCategoryDisplay(), "Laboratory report"));
   }
 
   /**
@@ -181,9 +223,9 @@ public class NotificationLaboratoryDataBuilder
    */
   @Deprecated(since = "1.2.1")
   private void setExampleType() {
-    typeSystem = requireNonNullElse(typeSystem, "http://loinc.org");
-    typeCode = requireNonNullElse(typeCode, "34782-3");
-    typeDisplay = requireNonNullElse(typeDisplay, "Infectious disease Note");
+    setTypeSystem(requireNonNullElse(getTypeSystem(), "http://loinc.org"));
+    setTypeCode(requireNonNullElse(getTypeCode(), "34782-3"));
+    setTypeDisplay(requireNonNullElse(getTypeDisplay(), "Infectious disease Note"));
   }
 
   /**
@@ -212,7 +254,8 @@ public class NotificationLaboratoryDataBuilder
    */
   @Deprecated(since = "1.2.1")
   private void setExampleCompositionStatus() {
-    compositionStatus = requireNonNullElse(compositionStatus, Composition.CompositionStatus.FINAL);
+    setCompositionStatus(
+        requireNonNullElse(getCompositionStatus(), Composition.CompositionStatus.FINAL));
   }
 
   /**
