@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.laboratory;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,30 +19,42 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
-
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.infectious.laboratory;
 
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.LOINC_ORG_SYSTEM;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.PATHOGEN_DETECTION_BASE_URL;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.SYSTEM_SNOMED;
+import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.ReferenceUtils.internalReference;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.Utils.generateUuidString;
 import static java.util.Objects.requireNonNullElse;
 
+import javax.annotation.CheckForNull;
 import lombok.Setter;
-import org.hl7.fhir.r4.model.*;
+import org.hl7.fhir.r4.model.Annotation;
+import org.hl7.fhir.r4.model.CodeableConcept;
+import org.hl7.fhir.r4.model.Coding;
+import org.hl7.fhir.r4.model.Meta;
+import org.hl7.fhir.r4.model.Observation;
+import org.hl7.fhir.r4.model.Patient;
+import org.hl7.fhir.r4.model.Quantity;
+import org.hl7.fhir.r4.model.Reference;
+import org.hl7.fhir.r4.model.Specimen;
+import org.hl7.fhir.r4.model.StringType;
+import org.hl7.fhir.r4.model.Type;
 
 @Setter
 public class PathogenDetectionDataBuilder {
 
   public static final String LABORATORY_CATEGORY_CODE = "laboratory";
-  public static final String LABORATORY_CATEGROY_SYSTEM =
+  public static final String LABORATORY_CATEGORY_SYSTEM =
       "http://terminology.hl7.org/CodeSystem/observation-category";
   public static final String LABORATORY_CATEGORY_DISPLAY = "Laboratory";
   public static final String OBSERVATION_INTERPRETATION_SYSTEM =
       "http://terminology.hl7.org/CodeSystem/v3-ObservationInterpretation";
-  Patient notifiedPerson;
-  Specimen specimen;
+
+  @CheckForNull private Patient notifiedPerson;
+  @CheckForNull private Specimen specimen;
   private String pathogenDetectionId;
   private String observationCodeSystem;
   private String observationCodeCode;
@@ -57,10 +74,70 @@ public class PathogenDetectionDataBuilder {
   private Observation.ObservationStatus status;
   private Type value;
   private String profileUrl;
+  private Annotation note;
+
+  public static Observation deepCopy(
+      final Observation observation, final Patient subject, final Specimen specimen) {
+    final PathogenDetectionDataBuilder resultBuilder = new PathogenDetectionDataBuilder();
+    resultBuilder
+        .setProfileUrl(observation.getMeta().getProfile().getFirst().getValueAsString())
+        .setStatus(observation.getStatus())
+        .setNotifiedPerson(subject)
+        .setSpecimen(specimen)
+        .setPathogenDetectionId(observation.getId());
+
+    final Coding category = observation.getCategoryFirstRep().getCodingFirstRep();
+    resultBuilder
+        .setCategoryCode(category.getCode())
+        .setCategoryDisplay(category.getDisplay())
+        .setCategorySystem(category.getSystem());
+
+    final Coding code = observation.getCode().getCodingFirstRep();
+    resultBuilder
+        .setObservationCodeCode(code.getCode())
+        .setObservationCodeDisplay(code.getDisplay())
+        .setObservationCodeSystem(code.getSystem());
+
+    // value
+    final Type copiedValue =
+        switch (observation.getValue()) {
+          case Quantity quantity ->
+              new Quantity().setValue(quantity.getValue()).setUnit(quantity.getUnit());
+          case CodeableConcept codeableConcept ->
+              new CodeableConcept().addCoding(codeableConcept.getCodingFirstRep());
+          case StringType stringType -> new StringType(stringType.getValue());
+          case null, default -> null;
+        };
+    if (copiedValue != null) {
+      resultBuilder.setValue(copiedValue);
+    }
+
+    final Coding interpretation = observation.getInterpretationFirstRep().getCodingFirstRep();
+    resultBuilder
+        .setInterpretationCode(interpretation.getCode())
+        .setInterpretationDisplay(interpretation.getDisplay())
+        .setInterpretationSystem(interpretation.getSystem());
+
+    // note
+    if (!observation.getNote().isEmpty()) {
+      resultBuilder.setNote(new Annotation().setText(observation.getNoteFirstRep().getText()));
+    }
+
+    // method
+    if (observation.hasMethod() && observation.getMethod().hasCoding()) {
+      final Coding codingFirstRep = observation.getMethod().getCodingFirstRep();
+      resultBuilder
+          .setMethodCode(codingFirstRep.getCode())
+          .setMethodDisplay(codingFirstRep.getDisplay())
+          .setMethodSystem(codingFirstRep.getSystem());
+    }
+
+    return resultBuilder.build();
+  }
 
   public PathogenDetectionDataBuilder setDefaultData() {
     categoryCode = LABORATORY_CATEGORY_CODE;
-    categorySystem = LABORATORY_CATEGROY_SYSTEM;
+    categorySystem = LABORATORY_CATEGORY_SYSTEM;
     categoryDisplay = LABORATORY_CATEGORY_DISPLAY;
     observationCodeSystem = LOINC_ORG_SYSTEM;
     pathogenDetectionId = generateUuidString();
@@ -113,8 +190,12 @@ public class PathogenDetectionDataBuilder {
     Observation observation = new Observation();
     observation.setStatus(status);
 
-    observation.setSubject(new Reference(notifiedPerson));
-    observation.setSpecimen(new Reference(specimen));
+    if (notifiedPerson != null) {
+      observation.setSubject(internalReference(notifiedPerson));
+    }
+    if (specimen != null) {
+      observation.setSpecimen(internalReference(specimen));
+    }
     observation.setId(pathogenDetectionId);
     observation.setStatus(status);
 
@@ -137,6 +218,8 @@ public class PathogenDetectionDataBuilder {
     CodeableConcept codeableConceptMethod = new CodeableConcept();
     codeableConceptMethod.addCoding(new Coding(methodSystem, methodCode, methodDisplay));
     observation.setMethod(codeableConceptMethod);
+
+    observation.addNote(note);
 
     return observation;
   }
@@ -202,11 +285,9 @@ public class PathogenDetectionDataBuilder {
 
   @Deprecated(since = "1.2.1")
   private void checkAndSetCategoryCode() {
-    categorySystem =
-        requireNonNullElse(
-            categorySystem, "http://terminology.hl7.org/CodeSystem/observation-category");
-    categoryCode = requireNonNullElse(categoryCode, "laboratory");
-    categoryDisplay = requireNonNullElse(categoryDisplay, "Laboratory");
+    categorySystem = requireNonNullElse(categorySystem, LABORATORY_CATEGORY_SYSTEM);
+    categoryCode = requireNonNullElse(categoryCode, LABORATORY_CATEGORY_CODE);
+    categoryDisplay = requireNonNullElse(categoryDisplay, LABORATORY_CATEGORY_DISPLAY);
   }
 
   @Deprecated(since = "1.2.1")
@@ -222,9 +303,7 @@ public class PathogenDetectionDataBuilder {
   @Deprecated(since = "1.2.1")
   private void checkAndSetInterpretationCode() {
     interpretationSystem =
-        requireNonNullElse(
-            interpretationSystem,
-            "http://terminology.hl7.org/CodeSystem/v3-ObservationInterpretation");
+        requireNonNullElse(interpretationSystem, OBSERVATION_INTERPRETATION_SYSTEM);
     interpretationCode = requireNonNullElse(interpretationCode, "POS");
     interpretationDisplay = requireNonNullElse(interpretationDisplay, "Positive");
   }
