@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.receipt;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,9 +19,8 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
-
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.receipt;
 
 import de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals.TelecomDataBuilder;
 import org.hl7.fhir.r4.model.ContactPoint;
@@ -24,17 +28,21 @@ import org.hl7.fhir.r4.model.Organization;
 
 public class DemisOrganizationBuilder {
 
+  private static final String ORGANIZATION_ID = "DEMIS";
+  private static final String ORGANIZATION_NAME = "DEMIS";
+  private static final String TELECOM_EMAIL = "demis-support@rki.de";
+
   public Organization createDemisOrg() {
     Organization demisOrg = new Organization();
-    demisOrg.setId("DEMIS");
-    demisOrg.setName("DEMIS");
+    demisOrg.setId(ORGANIZATION_ID);
+    demisOrg.setName(ORGANIZATION_NAME);
     Organization.OrganizationContactComponent demisOrgContact =
         new Organization.OrganizationContactComponent();
     demisOrgContact.addTelecom(
         new TelecomDataBuilder()
             .setSystem(ContactPoint.ContactPointSystem.EMAIL)
-            .setValue("demis@rki.de")
-            .buildContactPoint());
+            .setValue(TELECOM_EMAIL)
+            .build());
     demisOrg.addContact(demisOrgContact);
     return demisOrg;
   }
