@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.receipt;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,9 +19,8 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
-
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.receipt;
 
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.DEMIS_RKI_DE_FHIR;
 import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.NAMING_SYSTEM_NOTIFICATION_BUNDLE_ID;
@@ -38,6 +42,21 @@ import org.hl7.fhir.r4.model.Reference;
 /** This builder helps with the creation of receipt bundles. it will add all profile */
 @Data
 public class ReceiptBuilder {
+
+  private static final String RECEIPT_BUNDLE_PROFILE =
+      "https://demis.rki.de/fhir/StructureDefinition/ReceiptBundle";
+  private static final String NOTIFICATION_RECEIPT_PROFILE =
+      "https://demis.rki.de/fhir/StructureDefinition/NotificationReceipt";
+  private static final String RECEIVED_NOTIFICATION_EXTENSION =
+      "https://demis.rki.de/fhir/StructureDefinition/ReceivedNotification";
+  private static final String NOTIFICATION_ID_SYSTEM =
+      "https://demis.rki.de/fhir/NamingSystem/NotificationId";
+  private static final String REPORT_TYPE_CODE = "80563-0";
+  private static final String REPORT_TYPE_DISPLAY = "Report";
+  private static final String REPORT_TYPE_SYSTEM = "http://loinc.org";
+  private static final String PDF_CONTENT_TYPE = "application/pdf";
+  private static final String PDF_QUITTUNG_TITLE = "PDF Quittung";
+  private static final String RECEIVER_TITLE = "Empfänger";
 
   private Organization rkiOrganization;
 
@@ -63,8 +82,7 @@ public class ReceiptBuilder {
     Bundle receiptBundle = new Bundle();
 
     //    Bundle specific
-    receiptBundle.setMeta(
-        new Meta().addProfile("https://demis.rki.de/fhir/StructureDefinition/ReceiptBundle"));
+    receiptBundle.setMeta(new Meta().addProfile(RECEIPT_BUNDLE_PROFILE));
     receiptBundle.setType(Bundle.BundleType.COLLECTION);
 
     // Composition
@@ -72,22 +90,21 @@ public class ReceiptBuilder {
     notificationReceiptComposition.setId(generateUuidString());
     notificationReceiptComposition.setTitle("Meldevorgangsquittung");
 
-    notificationReceiptComposition.setMeta(
-        new Meta().addProfile("https://demis.rki.de/fhir/StructureDefinition/NotificationReceipt"));
+    notificationReceiptComposition.setMeta(new Meta().addProfile(NOTIFICATION_RECEIPT_PROFILE));
 
     Identifier value = new Identifier();
     value.setSystem(NAMING_SYSTEM_NOTIFICATION_BUNDLE_ID);
     value.setValue(notificationBundleId);
 
     Extension extension = new Extension();
-    extension.setUrl("https://demis.rki.de/fhir/StructureDefinition/ReceivedNotification");
+    extension.setUrl(RECEIVED_NOTIFICATION_EXTENSION);
     extension.setValue(value);
     notificationReceiptComposition.addExtension(extension);
 
     notificationReceiptComposition.setStatus(Composition.CompositionStatus.FINAL);
 
     notificationReceiptComposition.setType(
-        new CodeableConcept(new Coding("http://loinc.org", "80563-0", "Report")));
+        new CodeableConcept(new Coding(REPORT_TYPE_SYSTEM, REPORT_TYPE_CODE, REPORT_TYPE_DISPLAY)));
     notificationReceiptComposition.setDate(getCurrentDate());
 
     notificationReceiptComposition.addRelatesTo(createRelatesToTargetData());
@@ -126,9 +143,7 @@ public class ReceiptBuilder {
 
     relatesToTarget.setType("Composition");
     relatesToTarget.setIdentifier(
-        new Identifier()
-            .setSystem("https://demis.rki.de/fhir/NamingSystem/NotificationId")
-            .setValue(relatesToId));
+        new Identifier().setSystem(NOTIFICATION_ID_SYSTEM).setValue(relatesToId));
 
     component.setTarget(relatesToTarget);
     return component;
@@ -137,8 +152,8 @@ public class ReceiptBuilder {
   private void addRkiOrgWhenSet() {
     if (rkiOrganization != null) {
       Composition.SectionComponent section = new Composition.SectionComponent();
-      section.setTitle("Empfänger");
-      section.setCode(new CodeableConcept().setText("Empfänger"));
+      section.setTitle(RECEIVER_TITLE);
+      section.setCode(new CodeableConcept().setText(RECEIVER_TITLE));
       Reference receiverRef = new Reference(rkiOrganization);
       section.addEntry(receiverRef);
       notificationReceiptComposition.addSection(section);
@@ -155,10 +170,10 @@ public class ReceiptBuilder {
   private void addPdf(Binary pdfBinary, Bundle receiptBundle) {
 
     pdfBinary.setId(generateUuidString());
-    pdfBinary.setContentType("application/pdf");
+    pdfBinary.setContentType(PDF_CONTENT_TYPE);
     Composition.SectionComponent t1 = new Composition.SectionComponent();
-    t1.setTitle("PDF Quittung");
-    t1.setCode(new CodeableConcept().setText("PDF Quittung"));
+    t1.setTitle(PDF_QUITTUNG_TITLE);
+    t1.setCode(new CodeableConcept().setText(PDF_QUITTUNG_TITLE));
     Reference receiptBinaryRef = new Reference(pdfBinary);
     t1.addEntry(receiptBinaryRef);
     notificationReceiptComposition.addSection(t1);
