@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,32 +19,36 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
-
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals;
 
 import static java.util.Objects.requireNonNullElse;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
+import com.google.common.base.Strings;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+import java.util.SequencedCollection;
 import lombok.AccessLevel;
 import lombok.Setter;
+import org.apache.commons.lang3.tuple.Pair;
 import org.hl7.fhir.r4.model.Address;
 import org.hl7.fhir.r4.model.Coding;
 import org.hl7.fhir.r4.model.Extension;
 import org.hl7.fhir.r4.model.Organization;
 import org.hl7.fhir.r4.model.Reference;
-import org.hl7.fhir.r4.model.StringType;
 
 @Setter
-public class AddressDataBuilder {
+public class AddressDataBuilder implements FhirObjectBuilder {
 
   public static final String CURRENT = "current";
   public static final String ORDINARY = "ordinary";
   public static final String PRIMARY = "primary";
 
+  /** street or line information */
   private String street;
+
   private String houseNumber;
   private String additionalInfo;
   private String city;
@@ -49,6 +58,44 @@ public class AddressDataBuilder {
   @Setter(AccessLevel.PRIVATE)
   private List<Extension> extensionList = new ArrayList<>();
 
+  public static void copyExtensions(final Address source, final Address target) {
+    target.setExtension(source.getExtension().stream().map(Extension::copy).toList());
+  }
+
+  /**
+   * Create a new {@link Address} and only copy the first three letters of the postal code. See
+   * {@link AddressDataBuilder#copyExtensions(Address, Address)} on how to copy extensions.
+   */
+  public static Address copyOnlyPostalCode(final Address original) {
+    final Address result = new Address();
+    final String postalCode = original.getPostalCode();
+    if (!Strings.isNullOrEmpty(postalCode)) {
+      result.setPostalCode(postalCode.substring(0, Math.min(postalCode.length(), 3)));
+    }
+
+    return result;
+  }
+
+  /**
+   * Copy the given addresses so that they are redacted. Redacted means: all address types are
+   * copied, but only 3 letters of the postal code remain.
+   */
+  public static SequencedCollection<Address> copyOfRedactedAddress(
+      final Collection<Address> original) {
+    return original.stream()
+        .map(address -> Pair.of(address, AddressDataBuilder.copyOnlyPostalCode(address)))
+        .filter(
+            p ->
+                p.getRight()
+                    .hasPostalCode()) // Don't need addresses without postal code, they'd be empty
+        .map(
+            p -> {
+              AddressDataBuilder.copyExtensions(p.getLeft(), p.getRight());
+              return p.getRight();
+            })
+        .toList();
+  }
+
   /**
    * Standard builder method. All fields can be set to other input. Standard address: Dorfstraße 11
    * 21481 Buchhorst 20422
@@ -56,57 +103,21 @@ public class AddressDataBuilder {
    * @return hl7 address ressource to be used in address fields of other ressources
    */
   public Address buildExampleAddress() {
-
     street = requireNonNullElse(street, "Dorfstraße 11");
     city = requireNonNullElse(city, "Buchhorst");
     postalCode = requireNonNullElse(postalCode, "21481");
     country = requireNonNullElse(country, "20422");
-
     return build();
   }
 
-  /**
-   * @deprecated method. Use build() instead. Renaming to build() to be consistent with other
-   *     builders.
-   * @return
-   */
-  @Deprecated(since = "1.2.1")
-  public Address buildAddress() {
-    return build();
-  }
-
+  @Override
   public Address build() {
     final Address fhirAddress = new Address();
-    if (isNotBlank(postalCode)) {
-      fhirAddress.setPostalCode(postalCode);
-    }
-    if (isNotBlank(country)) {
-      fhirAddress.setCountry(country);
-    }
-
-    if (isNotBlank(street)) {
-
-      StringBuilder lineBuilder = new StringBuilder();
-      lineBuilder.append(street);
-
-      if (isNotBlank(houseNumber)) {
-        lineBuilder.append(" ").append(houseNumber);
-      }
-
-      if (isNotBlank(additionalInfo)) {
-        lineBuilder.append(" ").append(additionalInfo);
-      }
-
-      final StringType lineElement = fhirAddress.addLineElement();
-      lineElement.setValue(lineBuilder.toString());
-    }
-
-    if (isNotBlank(city)) {
-      fhirAddress.setCity(city);
-    }
-
+    addPostalCode(fhirAddress);
+    addCountry(fhirAddress);
+    addLine(fhirAddress);
+    addCity(fhirAddress);
     extensionList.forEach(fhirAddress::addExtension);
-
     return fhirAddress;
   }
 
@@ -134,5 +145,37 @@ public class AddressDataBuilder {
   public AddressDataBuilder addExtension(Extension extension) {
     extensionList.add(extension);
     return this;
+  }
+
+  private void addPostalCode(Address fhirAddress) {
+    if (isNotBlank(postalCode)) {
+      fhirAddress.setPostalCode(postalCode);
+    }
+  }
+
+  private void addCountry(Address fhirAddress) {
+    if (isNotBlank(country)) {
+      fhirAddress.setCountry(country);
+    }
+  }
+
+  private void addCity(Address fhirAddress) {
+    if (isNotBlank(city)) {
+      fhirAddress.setCity(city);
+    }
+  }
+
+  private void addLine(Address fhirAddress) {
+    if (isNotBlank(street)) {
+      StringBuilder lineBuilder = new StringBuilder();
+      lineBuilder.append(street);
+      if (isNotBlank(houseNumber)) {
+        lineBuilder.append(" ").append(houseNumber);
+      }
+      if (isNotBlank(additionalInfo)) {
+        lineBuilder.append(" ").append(additionalInfo);
+      }
+      fhirAddress.addLineElement().setValue(lineBuilder.toString());
+    }
   }
 }
