@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,9 +19,8 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
-
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals;
 
 import static java.util.Objects.requireNonNullElse;
 
@@ -24,46 +28,47 @@ import lombok.Setter;
 import org.hl7.fhir.r4.model.ContactPoint;
 
 @Setter
-public class TelecomDataBuilder {
+public class TelecomDataBuilder implements FhirObjectBuilder {
 
-  private String value;
   private ContactPoint.ContactPointSystem system;
   private ContactPoint.ContactPointUse use;
+  private String value;
 
   public ContactPoint buildExampleContactPoint() {
-
     value = requireNonNullElse(value, "987654320");
-    system = requireNonNullElse(system, ContactPoint.ContactPointSystem.FAX);
-    use = requireNonNullElse(use, ContactPoint.ContactPointUse.HOME);
-
-    return buildContactPoint();
+    system = requireNonNullElse(system, ContactPoint.ContactPointSystem.PHONE);
+    use = requireNonNullElse(use, ContactPoint.ContactPointUse.WORK);
+    return build();
   }
 
-  public ContactPoint buildContactPoint() {
-    ContactPoint newContactPoint = new ContactPoint();
-    if (value != null) {
-      newContactPoint.setValue(value);
-    }
-    if (system != null) {
-      newContactPoint.setSystem(system);
-    }
-    if (use != null) {
-      newContactPoint.setUse(use);
-    }
-    return newContactPoint;
+  public TelecomDataBuilder setEmail(String email) {
+    setSystem(ContactPoint.ContactPointSystem.EMAIL);
+    setValue(email);
+    return this;
+  }
+
+  public TelecomDataBuilder setPhone(String phone) {
+    setSystem(ContactPoint.ContactPointSystem.PHONE);
+    setValue(phone);
+    return this;
+  }
+
+  @Override
+  public ContactPoint build() {
+    return new ContactPoint().setValue(value).setSystem(system).setUse(use);
   }
 
   public ContactPoint buildPhoneContactPoint() {
     value = requireNonNullElse(value, "0123456789");
     system = ContactPoint.ContactPointSystem.PHONE;
     use = requireNonNullElse(use, ContactPoint.ContactPointUse.WORK);
-    return buildExampleContactPoint();
+    return build();
   }
 
   public ContactPoint buildEmailContactPoint() {
     value = requireNonNullElse(value, "some@email.com");
     system = ContactPoint.ContactPointSystem.EMAIL;
     use = requireNonNullElse(use, ContactPoint.ContactPointUse.WORK);
-    return buildExampleContactPoint();
+    return build();
   }
 }
