@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.receipt;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,13 +19,11 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
 
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.receipt;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import ca.uhn.fhir.context.FhirContext;
+import com.google.common.io.Resources;
+import de.gematik.demis.notification.builder.demis.fhir.notification.test.FhirJsonTestsUtil;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import org.hl7.fhir.r4.model.Organization;
@@ -30,17 +33,9 @@ class RkiOrganizationBuilderTest {
 
   @Test
   void shouldCreateRKIOrganization() throws IOException {
-
-    Organization demisOrg = new RkiOrganizationBuilder().createRkiOrganization();
-
-    String actualJson =
-        FhirContext.forR4().newJsonParser().setPrettyPrint(true).encodeResourceToString(demisOrg);
-
+    Organization organization = new RkiOrganizationBuilder().createRkiOrganization();
     String expectedJson =
-        new String(
-            getClass().getClassLoader().getResourceAsStream("RKIOrganization.json").readAllBytes(),
-            StandardCharsets.UTF_8);
-
-    assertThat(actualJson).isEqualToIgnoringNewLines(expectedJson);
+        Resources.toString(Resources.getResource("RKIOrganization.json"), StandardCharsets.UTF_8);
+    FhirJsonTestsUtil.assertEqualJson(organization, expectedJson, "Built RKI organization as JSON");
   }
 }
