@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.reports;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,60 +19,119 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
 
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.reports;
-
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
+import de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.Date;
+import org.hl7.fhir.r4.model.Coding;
 import org.hl7.fhir.r4.model.Composition;
+import org.hl7.fhir.r4.model.Identifier;
 import org.hl7.fhir.r4.model.PractitionerRole;
 import org.hl7.fhir.r4.model.QuestionnaireResponse;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class ReportBedOccupancyDataBuilderTest {
 
-  private PractitionerRole notifierRole;
-  private QuestionnaireResponse statisticInformationBedOccupancy;
+  private final ReportBedOccupancyDataBuilder builder = new ReportBedOccupancyDataBuilder();
+  private final PractitionerRole notifierRole = new PractitionerRole();
+  private final QuestionnaireResponse statisticInformationBedOccupancy =
+      new QuestionnaireResponse();
 
-  @BeforeEach
-  void setUp() {
-    notifierRole = new PractitionerRole();
-    statisticInformationBedOccupancy = new QuestionnaireResponse();
+  @Test
+  void setDefaults_shouldSetValues() {
+    builder.setDefaults();
+    Composition composition = builder.build();
+    Coding type = composition.getType().getCodingFirstRep();
+    assertThat(type.getSystem()).as("type system").isEqualTo("http://loinc.org");
+    assertThat(type.getCode()).as("type code").isEqualTo("80563-0");
+    Coding category = composition.getCategoryFirstRep().getCodingFirstRep();
+    assertThat(category.getSystem())
+        .as("category system")
+        .isEqualTo("https://demis.rki.de/fhir/CodeSystem/reportCategory");
+    assertThat(category.getCode()).as("category code").isEqualTo("bedOccupancyReport");
+    assertThat(composition.getTitle())
+        .as("title")
+        .isEqualTo("Bericht (Krankenhausbettenbelegungsstatistik)");
+    assertThat(composition.getId()).as("ID").isNotNull();
+    assertThat(composition.getDate()).as("date").isNotNull();
+    assertThat(composition.getStatus()).as("status").isEqualTo(Composition.CompositionStatus.FINAL);
+    assertThat(composition.getIdentifier()).as("notification ID").isNotNull();
+  }
+
+  @Test
+  void setDefaults_shouldKeepValues() throws ParseException {
+    String id = "init-id";
+    String identifier = "init-identifier";
+    String title = "init-title";
+    String typeCode = "init-type-code";
+    String categoryCode = "init-category-code";
+    Date currentDate = new SimpleDateFormat("dd.MM.yyyy").parse("13.04.2011");
+    Composition.CompositionStatus status = Composition.CompositionStatus.PRELIMINARY;
+    Composition composition =
+        new ReportBedOccupancyDataBuilder()
+            .setId(id)
+            .setIdentifierAsNotificationId(identifier)
+            .setTitle(title)
+            .setType(new Coding().setCode(typeCode))
+            .setCategory(new Coding().setCode(categoryCode))
+            .setCurrentDate(currentDate)
+            .setStatusStandard(status)
+            .setDefaults()
+            .build();
+    assertThat(composition.getId()).isEqualTo(id);
+    assertThat(composition.getIdentifier().getValue()).isEqualTo(identifier);
+    assertThat(composition.getTitle()).isEqualTo(title);
+    assertThat(composition.getType().getCodingFirstRep().getCode()).isEqualTo(typeCode);
+    assertThat(composition.getCategoryFirstRep().getCodingFirstRep().getCode())
+        .isEqualTo(categoryCode);
+    assertThat(composition.getDate()).isEqualTo(currentDate);
+  }
+
+  @Test
+  void shouldSetNotificationId() {
+    builder.setIdentifierAsNotificationId("notificationId");
+    Composition composition = builder.build();
+    Identifier notificationId = composition.getIdentifier();
+    assertThat(notificationId.getSystem())
+        .as("notificationId system")
+        .isEqualTo(DemisConstants.NOTIFICATION_ID_SYSTEM);
+    assertThat(notificationId.getValue()).as("notificationId value").isEqualTo("notificationId");
   }
 
   @Test
   void shouldCreateCompositionWithGivenTypeData() {
 
-    ReportBedOccupancyDataBuilder reportBundleDataBuilder = new ReportBedOccupancyDataBuilder();
+    builder.setNotifierRole(this.notifierRole);
+    builder.setStatisticInformationBedOccupancy(this.statisticInformationBedOccupancy);
 
-    reportBundleDataBuilder.setTypeCode("typeCode");
-    reportBundleDataBuilder.setTypeSystem("typeSystem");
-    reportBundleDataBuilder.setTypeDisplay("typeDisplay");
+    builder.setType(
+        new Coding().setCode("typeCode").setSystem("typeSystem").setDisplay("typeDisplay"));
 
-    Composition composition =
-        reportBundleDataBuilder.buildReportBedOccupancy(
-            notifierRole, statisticInformationBedOccupancy);
+    Composition composition = builder.build();
 
-    assertThat(composition.getType().getCodingFirstRep().getCode()).isEqualTo("typeCode");
-    assertThat(composition.getType().getCodingFirstRep().getSystem()).isEqualTo("typeSystem");
-    assertThat(composition.getType().getCodingFirstRep().getDisplay()).isEqualTo("typeDisplay");
+    Coding code = composition.getType().getCodingFirstRep();
+    assertThat(code.getCode()).isEqualTo("typeCode");
+    assertThat(code.getSystem()).isEqualTo("typeSystem");
+    assertThat(code.getDisplay()).isEqualTo("typeDisplay");
   }
 
   @Test
   void shouldCreateCompositionWithGivenCategoryData() {
-    ReportBedOccupancyDataBuilder reportBundleDataBuilder = new ReportBedOccupancyDataBuilder();
+    builder.setNotifierRole(this.notifierRole);
+    builder.setStatisticInformationBedOccupancy(this.statisticInformationBedOccupancy);
 
-    reportBundleDataBuilder.setCategoryCode("categoryCode");
-    reportBundleDataBuilder.setCategorySystem("categorySystem");
-    reportBundleDataBuilder.setCategoryDisplay("categoryDisplay");
+    builder.setCategory(
+        new Coding()
+            .setCode("categoryCode")
+            .setSystem("categorySystem")
+            .setDisplay("categoryDisplay"));
 
-    Composition composition =
-        reportBundleDataBuilder.buildReportBedOccupancy(
-            notifierRole, statisticInformationBedOccupancy);
+    Composition composition = builder.build();
 
     assertThat(composition.getCategoryFirstRep().getCodingFirstRep().getCode())
         .isEqualTo("categoryCode");
@@ -79,14 +143,12 @@ class ReportBedOccupancyDataBuilderTest {
 
   @Test
   void shouldCreateCompositionWithGivenSubjectData() {
-    ReportBedOccupancyDataBuilder reportBundleDataBuilder = new ReportBedOccupancyDataBuilder();
+    builder.setNotifierRole(this.notifierRole);
+    builder.setStatisticInformationBedOccupancy(this.statisticInformationBedOccupancy);
 
-    reportBundleDataBuilder.setSubjectSystem("subjectSystem");
-    reportBundleDataBuilder.setSubjectValue("subjectValue");
+    builder.setSubject(new Identifier().setSystem("subjectSystem").setValue("subjectValue"));
 
-    Composition composition =
-        reportBundleDataBuilder.buildReportBedOccupancy(
-            notifierRole, statisticInformationBedOccupancy);
+    Composition composition = builder.build();
 
     assertThat(composition.getSubject().getIdentifier().getSystem()).isEqualTo("subjectSystem");
     assertThat(composition.getSubject().getIdentifier().getValue()).isEqualTo("subjectValue");
@@ -94,14 +156,13 @@ class ReportBedOccupancyDataBuilderTest {
 
   @Test
   void shouldCreateCompositionWithGivenIdentifierData() {
-    ReportBedOccupancyDataBuilder reportBundleDataBuilder = new ReportBedOccupancyDataBuilder();
+    builder.setNotifierRole(this.notifierRole);
+    builder.setStatisticInformationBedOccupancy(this.statisticInformationBedOccupancy);
 
-    reportBundleDataBuilder.setIdentifierSystem("identifierSystem");
-    reportBundleDataBuilder.setIdentifierValue("identifierValue");
+    builder.setIdentifier(
+        new Identifier().setSystem("identifierSystem").setValue("identifierValue"));
 
-    Composition composition =
-        reportBundleDataBuilder.buildReportBedOccupancy(
-            notifierRole, statisticInformationBedOccupancy);
+    Composition composition = builder.build();
 
     assertThat(composition.getIdentifier().getSystem()).isEqualTo("identifierSystem");
     assertThat(composition.getIdentifier().getValue()).isEqualTo("identifierValue");
@@ -109,32 +170,48 @@ class ReportBedOccupancyDataBuilderTest {
 
   @Test
   void shouldCreateCompositionWithGivenStatusData() {
-    ReportBedOccupancyDataBuilder reportBundleDataBuilder = new ReportBedOccupancyDataBuilder();
+    builder.setNotifierRole(this.notifierRole);
+    builder.setStatisticInformationBedOccupancy(this.statisticInformationBedOccupancy);
 
-    reportBundleDataBuilder.setStatus("final");
+    builder.setStatus("final");
 
-    Composition composition =
-        reportBundleDataBuilder.buildReportBedOccupancy(
-            notifierRole, statisticInformationBedOccupancy);
+    Composition composition = builder.build();
 
     assertThat(composition.getStatus()).isEqualTo(Composition.CompositionStatus.FINAL);
   }
 
   @Test
   void shouldCreateCompositionWithGivenTitleIdAndDateData() {
-    ReportBedOccupancyDataBuilder reportBundleDataBuilder = new ReportBedOccupancyDataBuilder();
+    builder.setNotifierRole(this.notifierRole);
+    builder.setStatisticInformationBedOccupancy(this.statisticInformationBedOccupancy);
 
-    reportBundleDataBuilder.setTitle("title");
-    reportBundleDataBuilder.setId("id");
+    builder.setTitle("title");
+    builder.setId("id");
     Date date = new Date();
-    reportBundleDataBuilder.setCurrentDate(date);
+    builder.setCurrentDate(date);
 
-    Composition composition =
-        reportBundleDataBuilder.buildReportBedOccupancy(
-            notifierRole, statisticInformationBedOccupancy);
+    Composition composition = builder.build();
 
     assertThat(composition.getTitle()).isEqualTo("title");
     assertThat(composition.getId()).isEqualTo("id");
     assertThat(composition.getDate()).isEqualTo(date);
+  }
+
+  @Test
+  void shouldSetNotifierRole() {
+    builder.setNotifierRole(this.notifierRole);
+    Composition composition = builder.build();
+    assertThat(composition.getAuthor().getFirst().getResource())
+        .as("notifier role")
+        .isEqualTo(this.notifierRole);
+  }
+
+  @Test
+  void shouldSetStatistics() {
+    this.builder.setStatisticInformationBedOccupancy(this.statisticInformationBedOccupancy);
+    Composition composition = builder.build();
+    assertThat(composition.getSectionFirstRep().getEntryFirstRep().getResource())
+        .as("bed occupancy statistics")
+        .isEqualTo(this.statisticInformationBedOccupancy);
   }
 }
