@@ -1,6 +1,11 @@
-/*
- * Copyright [2023], gematik GmbH
- *
+package de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals;
+
+/*-
+ * #%L
+ * notification-builder-library
+ * %%
+ * Copyright (C) 2025 gematik GmbH
+ * %%
  * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the
  * European Commission – subsequent versions of the EUPL (the "Licence").
  * You may not use this work except in compliance with the Licence.
@@ -14,11 +19,11 @@
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ * #L%
  */
 
-package de.gematik.demis.notification.builder.demis.fhir.notification.builder.technicals;
-
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatNoException;
 
 import org.hl7.fhir.r4.model.Address;
 import org.hl7.fhir.r4.model.Coding;
@@ -101,5 +106,26 @@ class AddressDataBuilderTest {
   void shouldAddExtension() {
     Address address = new AddressDataBuilder().addExtension(new Extension()).build();
     assertThat(address.getExtension()).hasSize(1);
+  }
+
+  @Test
+  void thatCopyPostalCodeHandlesMissingPostalCode() {
+    final Address address = new Address().setPostalCode(null);
+    assertThatNoException().isThrownBy(() -> AddressDataBuilder.copyOnlyPostalCode(address));
+  }
+
+  @Test
+  void thatCopyPostalCodeHandlesVariousLengths() {
+    final Address address = new Address().setPostalCode("12345");
+    assertThat(AddressDataBuilder.copyOnlyPostalCode(address).getPostalCode()).isEqualTo("123");
+
+    address.setPostalCode("123");
+    assertThat(AddressDataBuilder.copyOnlyPostalCode(address).getPostalCode()).isEqualTo("123");
+
+    address.setPostalCode("12");
+    assertThat(AddressDataBuilder.copyOnlyPostalCode(address).getPostalCode()).isEqualTo("12");
+
+    address.setPostalCode("");
+    assertThat(AddressDataBuilder.copyOnlyPostalCode(address).getPostalCode()).isNull();
   }
 }
